+++ tests/test_class_docstrings.py
@@ -12,21 +12,25 @@ def test_simple_model(app, do_autodoc):
         "",
         "   :param id: Primary key: ID",
         "   :type id: ~django.db.models.AutoField",
-        "   :param dummy_field: Very verbose name of dummy field. This should help you",
+        "   :param dummy_field: Very verbose name of dummy field. "
+        "This should help you. Docstring of char field",
         "   :type dummy_field: ~django.db.models.CharField",
         "",
         "   Relationship fields:",
         "",
         "   :param file: File "
-        "(related name: :attr:`~dummy_django_app.models.FileModel.simple_models`)",
+        "(related name: :attr:`~dummy_django_app.models.FileModel.simple_models`). "
+        "Docstring of foreign key",
         "   :type file: :class:`~django.db.models.ForeignKey` to "
         ":class:`~dummy_django_app.models.FileModel`",
         "   :param childA: ChildA "
-        "(related name: :attr:`~dummy_django_app.models.ChildModelA.simple_model`)",
+        "(related name: :attr:`~dummy_django_app.models.ChildModelA.simple_model`). "
+        "Docstring of one to one field",
         "   :type childA: :class:`~django.db.models.OneToOneField` to "
         ":class:`~dummy_django_app.models.ChildModelA`",
         "   :param childrenB: ChildrenB "
-        "(related name: :attr:`~dummy_django_app.models.ChildModelB.simple_models`)",
+        "(related name: :attr:`~dummy_django_app.models.ChildModelB.simple_models`). "
+        "Docstring of many to many field",
         "   :type childrenB: :class:`~django.db.models.ManyToManyField` to "
         ":class:`~dummy_django_app.models.ChildModelB`",
         "",
@@ -59,21 +63,25 @@ def test_database_table(app, do_autodoc):
         "",
         "   :param id: Primary key: ID",
         "   :type id: ~django.db.models.AutoField",
-        "   :param dummy_field: Very verbose name of dummy field. This should help you",
+        "   :param dummy_field: Very verbose name of dummy field. "
+        "This should help you. Docstring of char field",
         "   :type dummy_field: ~django.db.models.CharField",
         "",
         "   Relationship fields:",
         "",
         "   :param file: File "
-        "(related name: :attr:`~dummy_django_app.models.FileModel.simple_models`)",
+        "(related name: :attr:`~dummy_django_app.models.FileModel.simple_models`). "
+        "Docstring of foreign key",
         "   :type file: :class:`~django.db.models.ForeignKey` to "
         ":class:`~dummy_django_app.models.FileModel`",
         "   :param childA: ChildA "
-        "(related name: :attr:`~dummy_django_app.models.ChildModelA.simple_model`)",
+        "(related name: :attr:`~dummy_django_app.models.ChildModelA.simple_model`). "
+        "Docstring of one to one field",
         "   :type childA: :class:`~django.db.models.OneToOneField` to "
         ":class:`~dummy_django_app.models.ChildModelA`",
         "   :param childrenB: ChildrenB "
-        "(related name: :attr:`~dummy_django_app.models.ChildModelB.simple_models`)",
+        "(related name: :attr:`~dummy_django_app.models.ChildModelB.simple_models`). "
+        "Docstring of many to many field",
         "   :type childrenB: :class:`~django.db.models.ManyToManyField` to "
         ":class:`~dummy_django_app.models.ChildModelB`",
         "",