+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 UNRELEASED
 ----------
 
+* Add inline docstrings of model fields to parameter documentation of models
 * Add support for Python 3.11
 * Add support for Django 4.1
 * Drop support for Django 2.2